refactor(scaffolding): extract single-test collaborator + shared renderer
Extract single-test scaffold orchestration into a new SingleTestScaffolder
collaborator and establish a shared TestScaffoldRenderer (target resolution,
constructor-argument synthesis, sibling-pattern extraction, per-framework
rendering) consumed directly by both the single-test flow and the
batch/first-test-file flows. ScaffoldingService is now a thin
IScaffoldingService facade for test scaffolding, mirroring the merged
TypeScaffolder extraction. Pure code motion — preview content, warnings,
tokens, logger categories, and the public facade contract are unchanged.

Deviation from plan (Directive #5 re-derivation vs live post-merge source):
ValidateIsTestProject and DetectTestFrameworkFromProjectFile are pinned by
string-based reflection in ScaffoldingFirstTestFileTests.cs (GetMethod on
typeof(ScaffoldingService), Instance|NonPublic), which find_references/grep
cannot see. Moving them into the static renderer would break those two tests
and force a 2nd test file into scope. They (and ResolveTestFramework) stay as
instance methods on the facade; the facade resolves project/framework before
delegating and threads the framework into the collaborator. This also makes
SingleTestScaffolder a pure (IWorkspaceManager, IFileOperationService)
collaborator with no unused logger, exactly mirroring TypeScaffolder.

Validated: dotnet build RoslynMcp.Roslyn + RoslynMcp.Tests (Release,
TreatWarningsAsErrors) clean; scoped regression ScaffoldingIntegrationTests +
ScaffoldingFirstTestFileTests 33/33 pass (incl. both reflection tests and the
repointed BuildArgExpression unit test).

Closes: scaffolding-single-test-collaborator-extraction
